--- 백준/Bronze/10818. 최소， 최대/최소， 최대.java
@@ -1,14 +1,0 @@
-import java.util.*;
-
-public class Main {
-    public static void main(String[] args) {
-        Scanner scanner = new Scanner(System.in);
-        int x = scanner.nextInt();
-        ArrayList<Integer> arr = new ArrayList<>();
-        for(int i = 0; i<x;i++){
-            arr.add(scanner.nextInt());
-        }
-        Collections.sort(arr);
-        System.out.printf("%d %d", arr.get(0), arr.get(x-1));
-    }
-}
